# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API SELETIVO",
        "version": "0.1"
    },
    "paths": {
...
        "/api/servidor-efetivo": {
            "get": {
                "tags": [
                    "Servidor-Efetivo"
                ],
                "summary": "GET Servidor Efetivo",
                "description": "GET Paginado",
                "operationId": "4b1c83d6265f7aad3977994940ab58c8",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page Number",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/store-servidor-efetivo": {
            "post": {
                "tags": [
                    "Servidor-Efetivo"
                ],
                "summary": "Vincula torna pessoa em servidor efetivo",
                "description": "Endpoint para vincular pessoa em servidor efetivo",
                "operationId": "7899fc2a52cda1f941915ce5c8159270",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pes_id",
                                    "se_matricula"
                                ],
                                "properties": {
                                    "pes_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "se_matricula": {
                                        "type": "string",
                                        "example": "205400"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vínculo realizado com sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Vínculo realizado com sucesso!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro na validação dos dados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro de validação"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/show-servidor-efetivo/{pes_id}": {
            "get": {
                "tags": [
                    "Servidor-Efetivo"
                ],
                "summary": "Show Servidor Efetivo",
                "description": "Show Servidor Efetivo",
                "operationId": "f5db8772bb347653f8b9f2c83e1c9195",
                "parameters": [
                    {
                        "name": "pes_id",
                        "in": "path",
                        "description": "ID da Pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Servidor Efetivo encontrado",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Nenhum vínculo de servidor efetivo encontrado para esta pessoa"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/update-servidor-efetivo/{pes_id}": {
            "put": {
                "tags": [
                    "Servidor-Efetivo"
                ],
                "summary": "Atualizar um servidor efetivo",
                "description": "Endpoint para atualizar servidor efetivo",
                "operationId": "83f3a6b1e8c141eda6eec890e50a497e",
                "parameters": [
                    {
                        "name": "pes_id",
                        "in": "path",
                        "description": "ID da Pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "se_matricula": {
                                        "type": "string",
                                        "example": "205400"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Servidor Efetivo Atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Servidor Efetivo Atualizado"
                                        },
                                        "data": {
                                            "properties": {
                                                "se_matricula": {
                                                    "type": "integer",
                                                    "example": "205400"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro na requisição"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/delete-servidor-efetivo/{pes_id}": {
            "delete": {
                "tags": [
                    "Servidor-Efetivo"
                ],
                "summary": "Deleta Servidor Efetivo",
                "description": "Remove Servidor Efetivo",
                "operationId": "3ad3cb7cfde5c883bf72d83ac0fa6e5c",
                "parameters": [
                    {
                        "name": "pes_id",
                        "in": "path",
                        "description": "ID da Pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Servidor Efetivo Removido",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Servidor Efetivo não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}